# Optimize string array buffer conversion