--- a/tiledb/npbuffer.h
+++ b/tiledb/npbuffer.h
@@ -1,0 +1,13 @@
+#include <pybind11/pybind11.h>
+#include <pybind11/numpy.h>
+#include <pybind11/pytypes.h>
+#include "util.h"
+
+namespace tiledbpy {
+
+namespace py = pybind11;
+using namespace pybind11::literals;
+
+py::tuple convert_np(py::array input, bool allow_unicode);
+
+}