Benchmark converting 10,000,000 random strings (length=10)

Current implementation:
```
CPU times: user 10.6 s, sys: 184 ms, total: 10.8 s
Wall time: 11.1 s
```

This PR:
```
CPU times: user 267 ms, sys: 52 ms, total: 319 ms
Wall time: 326 ms
```